+++ src/de/eppelt/roland/dominion/task/KerkermeisterOpfer.java
@@ -11,6 +11,9 @@ import de.eppelt.roland.dominion.ui.Handler;
 public class KerkermeisterOpfer extends OpferAufgabeImpl {
 	
 	
+	public static final String EINEN_FLUCH_AUF_DIE_HAND_NEHMEN = "Einen Fluch auf die Hand nehmen";
+
+
 	public static class FluchNehmen extends HexenOpfer {
 
 		public FluchNehmen(Spieler täter, Spieler opfer) {
@@ -31,6 +34,11 @@ public class KerkermeisterOpfer extends OpferAufgabeImpl {
 	}
 	
 	
+	public static enum Wahl { KEINE, FLUCH, ABLEGEN };
+	
+	
+	Wahl wahl = Wahl.KEINE;
+	int ablegen = 2;
 	int abgelegt = -1; // d.h. entscheiden
 
 
@@ -39,23 +47,47 @@ public class KerkermeisterOpfer extends OpferAufgabeImpl {
 	}
 	
 	
+	public String getMessage() {
+		switch (wahl) {
+			case KEINE:
+				return getOpfer().getName()+" prüft noch seine Optionen.";
+			case FLUCH:
+				return getOpfer().getName()+" hat einen Fluch genommen.";
+			case ABLEGEN:
+				return getOpfer().getName()+" hat "+karteN(ablegen)+" abgelegt.";
+			default:
+				return "Wahl "+wahl.name()+" unbekannt. Bitte den Entwickler infomieren.";
+		}
+	}
+
+	
+	@Override public void vorbereiten() {
+		super.vorbereiten();
+		ablegen = Math.min(2, getOpfer().getHandkarten().size());
+	}
+	
+	
 	@Override public boolean anzeigen() {
-		if (abgelegt>=2) {
+		if (abgelegt>=ablegen) {
 			done();
 			return false;
 		} else {
 			headerHandkartenTitle();
 			if (abgelegt<0) {
 				sayln("Entscheide dich:");
-				button("2 Karten ablegen", 'a', true, handler -> {
+				button(karteN(ablegen)+" ablegen", 'a', true, handler -> {
+					wahl = Wahl.ABLEGEN;
 					abgelegt = 0;
+					getTäter().updateMe();
 				});
 				ln();
-				button("Einen Fluch auf die Hand nehmen", 'f', true, handler -> {
+				button(EINEN_FLUCH_AUF_DIE_HAND_NEHMEN, 'f', true, handler -> {
 					handler.sofortAufgabe(new FluchNehmen(getTäter(), getOpfer()));
+					wahl = Wahl.FLUCH;
+					getTäter().updateMe();
 					done();
 				});
-			} else if (abgelegt==0) {
+			} else if (abgelegt==ablegen-2) {
 				sayln("Lege zwei Handkarten ab. Wähle aus:");
 				String anyKey = any(handkarten(), null);
 				ln();
@@ -67,15 +99,22 @@ public class KerkermeisterOpfer extends OpferAufgabeImpl {
 					}
 					abgelegt += index.length;
 				});
-			} else if (abgelegt==1) {
-				sayln("Du musst noch eine Handkarte ab legen. Wähle aus:");
+			} else if (abgelegt==ablegen-1) {
+				sayln("Du musst noch eine Handkarte ablegen. Wähle aus:");
 				oneIndex(handkarten(), (handler, index) -> {
 					handler.seite().legeAb(handkarten().ziehe(index));
+					abgelegt++;
 					done();
 				});
 			}
 			return true;
 		}
 	}
+
+
+	private String karteN(int anzahl) {
+		return Integer.toString(anzahl)+ (anzahl==1 ? " Karte" : " Karten");
+	}
+
 
 }